% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{CoefTable}
\alias{CoefTable}
\title{Generate Table of Coefficients}
\usage{
CoefTable(fit, weather = FALSE, sim = FALSE, xtable = FALSE)
}
\arguments{
\item{fit}{The fit generated by getFits of the illegal immigrant data set.}

\item{weather}{Whether or not weather was included in the fitting process. Defaults to FALSE.}

\item{sim}{Is the fit from the simulation study or the empirical example? Defaults to FALSE for empirical example fit.}

\item{xtable}{Should results be returned in xtable format? Defaults to FALSE.}
}
\description{
This function takes the fits from either the empirical example or sim study and generates the table of coefficients and returns the parameter estimate, standard error, t-value and p-value.
}
\examples{
CoefTable()
}
\keyword{output}

