% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{DisplayEmpiricalOutput}
\alias{DisplayEmpiricalOutput}
\title{Display the Empirical Output results}
\usage{
DisplayEmpiricalOutput(fits)
}
\arguments{
\item{fits}{Fits from getFits()}
}
\description{
This function takes the fits generated by the empirical output analysis and displays the results.
}
\examples{
DisplayEmpiricalOutput()
}
\keyword{output}

