% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{DisplaySimOutput}
\alias{DisplaySimOutput}
\title{Display the Simulation Study results}
\usage{
DisplaySimOutput(N)
}
\arguments{
\item{N}{Total number of simulation iterations.}
}
\description{
This function takes the result files generated by the simulation study analysis and displays the results.
}
\examples{
DisplaySimOutput()
}
\keyword{output,}
\keyword{simulation}

