% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getBias}
\alias{getBias}
\title{Calculate bias statistic}
\usage{
getBias(est, true)
}
\arguments{
\item{est}{The parameter value estimated}

\item{true}{The parameter given in the simulation study design.}
}
\description{
Calculates the difference between the estimated parameter and the true parameter for the simulation study.
}
\examples{
getBias()
}
\keyword{bias}

