% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getDevianceResiduals}
\alias{getDevianceResiduals}
\title{Calculate Deviance Residuals}
\usage{
getDevianceResiduals(fits)
}
\arguments{
\item{fits}{}
}
\description{
Calculates the Deviance of the residuals
}
\examples{
getDevianceResiduals()
}
\keyword{deviance,}
\keyword{residuals}

