% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getEstimates}
\alias{getEstimates}
\title{General Simulation Function}
\usage{
getEstimates(N = 1000, Time = 365, beta = c(-7, -0.5, -0.25, 0.25, 0.5,
  0), sim = 25, method = 1, seed = 4885)
}
\arguments{
\item{N}{Number of people in the simulation. Defaults to 1000.}

\item{Time}{Total length of simulation in days. Defaults to 365 for one year.}

\item{beta}{True-values for parameter generation.}

\item{sim}{Number of simulation iterations. Defaults to 25 in case I accidentally run it.}

\item{method}{Use method one or two, where method 1 is random movement between locations and method 2 prefers remaining with less chance of switching.}

\item{seed}{Seed to use for deterministic reasons. Defaults to 4885 for my birthday.}
}
\description{
Calls simCat functions to generate data, then method functions to fill in TVCs, then estimates using optim, then calls SaveData() to file it. Basically the whole simulation script for convenience.
}
\examples{
getEstimates()
}
\keyword{script}

