% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getFits}
\alias{getFits}
\title{Estimate fits for empirical example}
\usage{
getFits(seed, beta, weather = FALSE)
}
\arguments{
\item{seed}{Starting seed}

\item{beta}{Beginning true values}

\item{weather}{Whether or not to include weather for the year in the calculations. Defaults to FALSE.}
}
\description{
Estimates the fits for the empirical example
}
\examples{
getFits()
}
\keyword{estimation}
\keyword{simulation,}

