% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getNHat}
\alias{getNHat}
\title{Population size estimation}
\usage{
getNHat(x.arr, ev.mat, b)
}
\arguments{
\item{x.arr}{3-dimensional array of covariates, time-varying or otherwise, with dim[1] representing persons, dim[2] representing covariates and dim[3] representing time.}

\item{ev.mat}{Matrix of ones and zeroes with N rows and t columns representing event occurrence (1) or non-occurrence (0) per day per person.}

\item{b}{Estimated beta coefficients from optim procedure.}
}
\description{
Generates estimate of population size as given by the Horvitz-Thompson estimation procedure
}
\examples{
getNHat()
}
\keyword{nhat}

