% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getPearsonChiSqR}
\alias{getPearsonChiSqR}
\title{Estimate pearson chi square residuals (in progress)}
\usage{
getPearsonChiSqR(fits)
}
\arguments{
\item{fits}{Fits for the simulation study}
}
\description{
This function isn't currently in use because I'm not sure whether or not this is a valid calculation.
}
\examples{
getPearsonChiSqR()
}
\keyword{residuals}

