% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getRMSE}
\alias{getRMSE}
\title{Calculate Root Mean Square Error}
\usage{
getRMSE(est, true)
}
\arguments{
\item{est}{The parameter value estimated}

\item{true}{The parameter given in the simulation study design.}
}
\description{
Calculates the root mean square error
}
\examples{
getRMSE()
}
\keyword{RMSE}

