% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{GetSimNhats}
\alias{GetSimNhats}
\title{Population Size Estimation for Simulation}
\usage{
GetSimNhats(nhats, N, xtable = FALSE)
}
\arguments{
\item{nhats}{These are the nhats from the simulation study.}

\item{N}{Total number of simulation iterations.}

\item{xtable}{Whether or not results should be formatted with xtable. Defaults to FALSE.}
}
\description{
Takes the population size estimations from the individual simulation study iterations and produces a table for display. Called from DisplaySimNhats.R
}
\examples{
GetSimNhats()
}
\keyword{output,}
\keyword{simulation}

