% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{GetSimPars}
\alias{GetSimPars}
\title{Beta Coefficient Estimates for Simulation}
\usage{
GetSimPars(pars, true.value, xtable = FALSE)
}
\arguments{
\item{pars}{Parameter (Beta) estimates from simulation study}

\item{true.value}{The true value of the parameters that were estimated in the sim study.}

\item{xtable}{Whether or not results should be formatted with xtable. Defaults to FALSE.}
}
\description{
Displays the beta estimates for the simulation study.
}
\examples{
GetSimPars()
}
\keyword{output,}
\keyword{simulation}

