% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{giveMeDataWithoutDetentions}
\alias{giveMeDataWithoutDetentions}
\title{Setup data from empirical example}
\usage{
giveMeDataWithoutDetentions(weather = FALSE)
}
\arguments{
\item{weather}{Whether or not to use the year's weather in the calculations. Defaults to FALSE.}
}
\description{
This function just exists for me to more easily get the data from my empirical example into a usable format.
}
\examples{
giveMeDataWithoutDetentions()
}
\keyword{data}
\keyword{import}

