% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{makeFirstValXhat}
\alias{makeFirstValXhat}
\title{Poisson Equivalent Method}
\usage{
makeFirstValXhat(x.arr, tvc.cols)
}
\arguments{
\item{x.arr}{3-dimensional array of covariates, time-varying or otherwise, with dim[1] representing persons, dim[2] representing covariates and dim[3] representing time.}

\item{tvc.cols}{Index of which columns vary over time.}
}
\description{
This is the function for the method deemed Poisson-Equivalent or "First-Value". Takes the N by Covariate by Time array and an index of which columns vary over time and produces an array with the missing time-varying covariates filled in with the first value that occurs.
}
\examples{
makeFirstValXhat()
}
\keyword{method,}
\keyword{tvc}

