% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{makeNvcbXhat}
\alias{makeNvcbXhat}
\title{Interval-stable Method}
\usage{
makeNvcbXhat(x.arr, ev)
}
\arguments{
\item{x.arr}{3-dimensional array of covariates, time-varying or otherwise, with dim[1] representing persons, dim[2] representing covariates and dim[3] representing time.}

\item{ev}{Event matrix of ones and zeroes with N rows and t columns representing event occurrence (1) or non-occurrence (0) per day per person.}
}
\description{
This is the function for the method deemed "Interval-Stable" method or "Next-value carried backwards." Takes the N by Covariate by Time array and an index of which columns vary over time and produces an array with the missing time-varying covariates filled in with the next value that is present chronologically. Fills in the last interval with the last value chronologically.
}
\examples{
makeNvcbXhat()
}
\keyword{method,}
\keyword{tvc}

