% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{makePropXhat}
\alias{makePropXhat}
\title{Proportional Representation Method}
\usage{
makePropXhat(x.arr, ev)
}
\arguments{
\item{x.arr}{3-dimensional array of covariates, time-varying or otherwise, with dim[1] representing persons, dim[2] representing covariates and dim[3] representing time.}

\item{ev}{Event matrix of ones and zeroes with N rows and t columns representing event occurrence (1) or non-occurrence (0) per day per person.}
}
\description{
This is the function for the method deemed "Proportional Representation".  Takes the N by Covariate by Time array and an index of which columns vary over time and produces an array with the missing time-varying covariates filled in sampling from all possible values obtained.
}
\examples{
makePropXhat()
}
\keyword{method,}
\keyword{tvc}

