% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{NHatTable}
\alias{NHatTable}
\title{Table of Population Size Estimates for Empirical Example}
\usage{
NHatTable(fits.w, xtable = FALSE)
}
\arguments{
\item{fits.w}{The fits from the empirical example including weather for the year.}

\item{xtable}{Whether or not results should be formatted with xtable. Defaults to FALSE.}
}
\description{
Generates a table of the population size estimates from the different methods on the empirical example.
}
\examples{
NHatTable()
}
\keyword{empirical}
\keyword{output,}

