% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{SaveData}
\alias{SaveData}
\title{Save Simulation Data}
\usage{
SaveData(data, sim.name)
}
\arguments{
\item{data}{List of all three methods and estimated coefficients followed by nhats.}

\item{sim.name}{Which method was used (1 or 2)}
}
\description{
Takes the data generated by the simulation study and saves it in files.
}
\examples{
SaveData()
}
\keyword{output,}
\keyword{simulation}

