% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{simCat}
\alias{simCat}
\title{Simulation Method 1 (Random movement)}
\usage{
simCat(N, Time, b)
}
\arguments{
\item{N}{Total number of people}

\item{Time}{Total length of simulation in days}

\item{b}{Covariate values for simulating risk of event.}
}
\description{
Time-varying categorical covariate representing where a person is on any given day. They have two primary locations - home and work with .65 and .3 probability respectively. There is also .05 probability to be randomly in a different city on any given day.
}
\examples{
simCat()
}
\keyword{simulation}

