% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{simCat2}
\alias{simCat2}
\title{Simulation Method 2 (Structured movement)}
\usage{
simCat2(N, Time, b)
}
\arguments{
\item{N}{Total number of people}

\item{Time}{Total length of simulation in days}

\item{b}{Covariate values for simulating risk of event.}
}
\description{
# Time-varying categorical covariate representing a value that has a 25% chance to change to a "safer" region, but only when there is an event.
}
\examples{
simCat2()
}
\keyword{simulation}

